update actions: for async actions
before the update, just TOGGLE_TODO action handles updating the
completed status and dispatched to the state. Therefore the logic of
changing the completed status and storing it was included in the
reducers.

But for the async dispatch to upload the save to the firebase, we use
thunk to return a function. So the startToggleTodo action first updates
the content to firebase(using ref.update(updates) method). And
then(es6 promise), it dispatch the revised action updateTodo(formerly
toggleTodo) to the state with the parameter id and updates.

Once the reducer gets its action dispatch, it no longer need to handle
the logic but instead it updates the passed in updates object by
overiding the previous todo. In this case a spread operator is used to
be able to override.

Voila!