update fetching data method: from firebase
Instead of fetching datas from localStorage, now all the todo datas are
fetched from firebase datbase.

Prvious local storage fetching methodi in the API (getState, setState)
is erased and its related tests are deleted.

Now in the app.jsx, when starting an app, a new action is dispatched to
store which fetches data from firebase.

This action first takes all the datab from firebase in its form.
Like { key : { bla1: bla1, bla2: bla2}, ...}. So to able to use this data in the store, it
has to be parsed as [ {id: key, bla1: bla1, bla2: bla2 ... }, ...]. Then
new array is pushed in to new array and dispatched to the store.

Lovely.